backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
pydantic>=2.7.0
numpy>=1.24.0
scikit-learn>=1.3.0
sentence-transformers>=2.6.1
python-dotenv>=1.0.1
google-generativeai>=0.8.3
requests>=2.32.3 